requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.6
django-environ==0.11.2
djangorestframework==3.14.0
idna==3.4
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
termcolor==2.3.0
timespan==0.2.0
urllib3==2.0.7
